Common course pairings frequencies
Calculates the frequency of each course pairing in the history of courses taken together.  The result gets appended to the excel document listing the pairings of the common classes.  This can be rearranged in Excel to get which classes are most frequently taken together.  See email for the excel document.